Catch failures properly

+++ scripts/install_snowboy.sh
@@ -10,10 +10,13 @@ SNOWBOY_VERSION="v1.1.0"
 SNOWBOY_REPOSITORY="https://github.com/Kitt-AI/snowboy.git"
 
 # Get the snowboy code
-if [ $(which git) ]; then
+if [ $(which git) ]
+then
   echo "Downloading snowboy."
-  git clone --branch $SNOWBOY_VERSION $SNOWBOY_REPOSITORY $TMP_DIR \
-    || { echo "Error: Failed to clone the snowboy repository."; exit 1 }
+  git clone --branch $SNOWBOY_VERSION $SNOWBOY_REPOSITORY $TMP_DIR
+  if [ ! $? -eq 0 ]; then
+    echo "Error: Failed to clone the snowboy repository."; exit 1
+  fi
 else
   echo "Error: Installing snowboy requires git."
   exit 1
@@ -41,7 +44,11 @@ fi
 # Make the module
 echo "Building..."
 cd $TMP_DIR/swig/Python
-make || { echo "Error: Build failed."; exit 1 }
+make
+if [ ! $? -eq 0 ]; then
+  echo "Error: Build failed."
+  exit 1
+fi
 cd $CURRENT_DIR
 
 # Create an init file for the module
@@ -49,8 +56,11 @@ touch $TMP_DIR/swig/Python/__init__.py
 
 # Move the module to the install directory
 echo "Installing module to $INSTALL_DIR/snowboydetect."
-mv $TMP_DIR/swig/Python $INSTALL_DIR/snowboydetect \
-  || { echo "Error: Failed to install module to $INSTALL_DIR/snowboydetect."; exit 1 }
+mv $TMP_DIR/swig/Python $INSTALL_DIR/snowboydetect
+if [ ! $? -eq 0 ]; then
+  echo "Error: Failed to install module to $INSTALL_DIR/snowboydetect."
+  exit 1
+fi
 
 # Clean up
 rm -r $TMP_DIR
